Mon, Jan 30, 2023, 9:08 PM -05:00

+++ kochan/10_character_strings/exercise_10/07_string_to_float.c
@@ -42,9 +42,9 @@ int main(void)
     for (; text[i] != '\0'; i++)
     {
 
-        int digit = text[i] - 48;
+        double digit = text[i] - 48;
 
-        post_decimal = (double)(post_decimal * 10 + digit);
+        post_decimal = (double)(post_decimal * 10.0 + digit);
 
         count++;
     }
